osint-app/pyproject.toml
[project]
name = "osint_app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "typeguard (>=4.4.2,<5.0.0)",
    "selenium (>=4.32.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "shodan (>=1.31.0,<2.0.0)",
    "wappalyzer (>=1.0.13,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "rich-click (>=1.8.9,<2.0.0)",
]

[tool.poetry]
packages = [{include = "osint_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
